% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_network_colnames.R
\name{check_network_colnames}
\alias{check_network_colnames}
\title{check_network_colnames()}
\usage{
check_network_colnames(networks)
}
\arguments{
\item{networks}{networks dataframe}
}
\value{
networks dataframe
}
\description{
Helper function that makes sure correct colnames are used. Will be removed eventually.
}
