% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_trials.R
\name{check_trials}
\alias{check_trials}
\title{check_trials()}
\usage{
check_trials(event_data, networks, ILV_tv = NULL, t_weights = NULL)
}
\arguments{
\item{event_data}{dataframe with columns id, trial, time, t_end}

\item{networks}{networks dataframe with columns trial, focal, other, and one or more columns of edge weights named descriptively.}

\item{ILV_tv}{dataframe of time-varying ILVs}

\item{t_weights}{dataframe of transmission weights, can be timevarying}
}
\value{
error message if trials are missing from dataframes
}
\description{
Helper function that ensures all trials are present in networks and events df
}
