% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_veff_type.R
\name{check_veff_type}
\alias{check_veff_type}
\title{check_veff_type}
\usage{
check_veff_type(veff_type)
}
\arguments{
\item{veff_type}{string or character vector user entered}
}
\value{
formatted veffs, ofc
}
\description{
check_veff_type
}
