% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_network_grid.R
\name{complete_network_grid}
\alias{complete_network_grid}
\title{pads out networks with all combinations of IDs and trials}
\usage{
complete_network_grid(event_data)
}
\arguments{
\item{event_data}{}
}
\value{
a dataframe containing all trials, dyads and times (but not inter-event intervals)
}
\description{
pads out networks with all combinations of IDs and trials
}
