% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_C_matrix.R
\name{create_C_matrix}
\alias{create_C_matrix}
\title{create_knowledge_matrix: private function to create the knowledge state matrix C\link{k,t,n}}
\usage{
create_C_matrix(event_data)
}
\arguments{
\item{event_data}{dataframe (event_data) in import_NBDA_STb.R / import_user_STb.R}
}
\value{
matrix \link{k,t,n} where value=0 if naive at t, otherwise 1
}
\description{
create_knowledge_matrix: private function to create the knowledge state matrix C\link{k,t,n}
}
