% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_knowledge_matrix.R
\name{create_knowledge_matrix}
\alias{create_knowledge_matrix}
\title{create_knowledge_matrix: private function to create the knowledge state matrix C\link{t,n}}
\usage{
create_knowledge_matrix(event_data)
}
\arguments{
\item{event_data}{dataframe (event_data) in import_NBDA_data.R / create_STb_data.R}
}
\value{
matrix \link{t,n} where value=0 if naive at t, otherwise 1
}
\description{
create_knowledge_matrix: private function to create the knowledge state matrix C\link{t,n}
}
