% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_W_matrix.R
\name{create_W_matrix}
\alias{create_W_matrix}
\title{create_W_matrix()}
\usage{
create_W_matrix(t_weights, max_time)
}
\arguments{
\item{t_weights}{dataframe with t_weight. all ids across all trials must be present.}

\item{max_time}{end of the observation period}
}
\value{
W dimensions k,t,n
}
\description{
Helper function to transform transmission weights into same size as Z
}
