% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_Z_matrix.R
\name{create_Z_matrix}
\alias{create_Z_matrix}
\title{create_Z_matrix: private function to create the knowledge state matrix Z\link{k,t,n}}
\usage{
create_Z_matrix(event_data)
}
\arguments{
\item{event_data}{dataframe (event_data) in import_NBDA_STb.R / import_user_STb.R}
}
\value{
matrix \link{k,t,n} where value=0 if naive at t, otherwise 1
}
\description{
create_Z_matrix: private function to create the knowledge state matrix Z\link{k,t,n}
}
