% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_ILV_datatype.R
\name{detect_ILV_datatype}
\alias{detect_ILV_datatype}
\title{helper function for importing ILVs}
\usage{
detect_ILV_datatype(x)
}
\arguments{
\item{x}{column or vector of data}
}
\description{
helper function for importing ILVs
}
