% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl_sanity_check.R
\name{dl_sanity_check}
\alias{dl_sanity_check}
\title{dl_sanity_check()}
\usage{
dl_sanity_check(data_list)
}
\arguments{
\item{data_list}{a datalist as generated by import_NBDA_STb or import_user_STb}
}
\value{
NULL, just print sanity check
}
\description{
Helper function to check data_list
}
