% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_bisonr_edgeweights.R
\name{extract_bisonr_edgeweights}
\alias{extract_bisonr_edgeweights}
\title{extract_bisonr_edgeweights}
\usage{
extract_bisonr_edgeweights(bison_fit, draws = 100)
}
\arguments{
\item{bison_fit}{object of class "bison_model"}

\item{draws}{number of draws from posterior, defaults to 100}
}
\value{
dataframe with cols draw, from, to, value
}
\description{
Creates networks dataframe for STb from bisonr fit with a given number of draws from the posterior distribution of edge weights.
}
