% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ILV.R
\name{extract_ILV}
\alias{extract_ILV}
\title{extract_ILV()}
\usage{
extract_ILV(nbda_object, ILV_type)
}
\arguments{
\item{nbda_object}{NBDAdata object}

\item{ILV_type}{character indicating types of ILVs: @asocILVdata = asocial, @intILVdata = social, and @multiILVdata = multiplicative (affects both social and asocial equally)}
}
\value{
List of vectors for each ILV
}
\description{
Helper function to extract the ILV matrix and back transform into vectors for STb format
}
