% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_LOOIC.R
\name{extract_LOOIC}
\alias{extract_LOOIC}
\title{extract_LOOIC()}
\usage{
extract_LOOIC(fit)
}
\arguments{
\item{fit}{A STb model fit}
}
\value{
LOOIC estimate
}
\description{
Convenience to quickly extract LOOIC scores. You should really be using \code{\link[=STb_compare]{STb_compare()}}.
}
