% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_strand_edgeweights.R
\name{extract_strand_edgeweights}
\alias{extract_strand_edgeweights}
\title{extract_bisonr_edgeweights}
\usage{
extract_strand_edgeweights(strand_obj, draws = 100)
}
\arguments{
\item{strand_obj}{object of class "STRAND Results Object"}

\item{draws}{number of draws from posterior, defaults to 100}
}
\value{
dataframe with cols draw, from, to, value
}
\description{
Creates networks dataframe for STb from STRAND results obj with a given number of draws from the posterior distribution of edge weights.
}
