% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_WAIC.R
\name{extract_WAIC}
\alias{extract_WAIC}
\title{extract_WAIC()}
\usage{
extract_WAIC(fit)
}
\arguments{
\item{fit}{A STb model fit}
}
\value{
WAIC estimate
}
\description{
Convenience function to extract WAIC scores. You should really be using \code{\link[=STb_compare]{STb_compare()}}.
}
