% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_STb_model.R
\name{fit_STb}
\alias{fit_STb}
\title{Fit STb model using cmdstanr}
\usage{
fit_STb(data_list, model_obj, ...)
}
\arguments{
\item{data_list}{list object exported from import_user_STb or import_NBDA_STb}

\item{model_obj}{Can be either model object exported from generate_STb_model or a filename}

\item{...}{Additional arguments passed to mod$sample (e.g., chains, iter_warmup, iter_sampling, adapt_delta)}
}
\value{
A CmdStanMCMC fit object
}
\description{
Fit STb model using cmdstanr
}
