% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_stancode.R
\name{format_stancode}
\alias{format_stancode}
\title{format_stancode}
\usage{
format_stancode(stan_code, indent = 4)
}
\arguments{
\item{stan_code}{string of stan code, duh}

\item{indent}{integer of num spaces for indent}
}
\value{
formatted string of stan code
}
\description{
format_stancode
}
