% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_STb_model_OADA.R
\name{generate_STb_model_OADA}
\alias{generate_STb_model_OADA}
\title{generate_STb_model_OADA()}
\usage{
generate_STb_model_OADA(
  STb_data,
  model_type = "full",
  transmission_func = "standard",
  separate_s = TRUE,
  veff_params = c(),
  veff_type = "id",
  gq = TRUE,
  priors = list()
)
}
\arguments{
\item{STb_data}{a list of formatted data returned from the STbayes_data() function}

\item{model_type}{string specifying the model type: "asocial" or "full"}

\item{transmission_func}{string specifying transmission function: "standard", "freqdep_f" or "freqdep_k" for frequency dependent complex contagion. Defaults to "standard".}

\item{veff_params}{Vector of parameter names (string) for which to estimate varying effects. For the moment, this is mutually exclusive for varying effects on trial or ID, you must choose one or the other using argument \code{veff_type}. Default is no varying effects.}

\item{veff_type}{string specifying whether varying effects should be applied to "ID" or "trial". Default is ID.}

\item{gq}{Boolean to indicate whether the generated quantities block is added (incl. ll for WAIC)}

\item{priors}{named list with strings containing the prior for log s or f.}
}
\value{
A STAN model (character) that is customized to the input data.
}
\description{
Helper function that creates stan code for OADA type models.
}
