% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_STb_model_TADA_sdirect.R
\name{generate_STb_model_TADA_sdirect}
\alias{generate_STb_model_TADA_sdirect}
\title{Dynamically generate STAN model based on input data}
\usage{
generate_STb_model_TADA_sdirect(
  STb_data,
  model_type = "full",
  intrinsic_rate = "constant",
  transmission_func = "standard",
  dTADA = F,
  veff_ID = c(),
  gq = TRUE,
  est_acqTime = FALSE,
  priors = list()
)
}
\arguments{
\item{STb_data}{a list of formatted data returned from the STbayes_data() function}

\item{model_type}{string specifying the model type: "asocial" or "full"}

\item{intrinsic_rate}{string specifying whether intrinsic rate is static or can change over time ("constant", "weibull")}

\item{transmission_func}{string specifying transmission function: "standard", "freqdep_f" or "freqdep_k" for complex contagion. Defaults to "standard". Complex contagion with multi-network model is not supported.}

\item{dTADA}{boolean indicating whether dTADA should be used.}

\item{veff_ID}{Parameters for which to estimate varying effects by individuals. Default is no varying effects.}

\item{gq}{Boolean to indicate whether the generated quantities block is added (incl. ll for WAIC)}

\item{est_acqTime}{Boolean to indicate whether gq block includes estimates for acquisition time. At the moment this uses 'one weird trick' to accomplish this and does not support estimates for non-integer learning times.}

\item{priors}{named list with strings containing the prior for log baserate, s, f, k.}
}
\value{
A STAN model (character) that is customized to the input data.
}
\description{
Dynamically generate STAN model based on input data
}
