% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ILVx_matrices.R
\name{generate_X_matrix}
\alias{generate_X_matrix}
\title{generate_X_matrix()}
\usage{
generate_X_matrix(ilv_vector, ilv_name, n_levels)
}
\arguments{
\item{ilv_vector}{c() of ILV values}

\item{ilv_name}{name to use for column}

\item{n_levels}{number of levels in factor}
}
\value{
matrix of size \verb{[N,L-1\]} where L=num levels

[N,L-1]: R:N,L-1\%5C
}
\description{
Helper function to create model matrices for categorical ILVs
}
