% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_network_term.R
\name{get_network_term}
\alias{get_network_term}
\title{Helper function to generate network terms for STAN}
\usage{
get_network_term(
  transmission_func,
  is_distribution = FALSE,
  separate_s = FALSE,
  num_networks = 1,
  veff_ID = c(),
  net_var = "A",
  net_index = "network",
  id_var = "id",
  trial_var = "trial",
  time_var = "time_step"
)
}
\description{
Helper function to generate network terms for STAN
}
