% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ST_prob_term.R
\name{get_ST_prob_term}
\alias{get_ST_prob_term}
\title{get_ST_prob_term()}
\usage{
get_ST_prob_term(
  transmission_func,
  is_distribution = FALSE,
  separate_s = FALSE,
  veff_ID = c(),
  num_networks = 1,
  s_var = "s_prime",
  id_var = "id",
  trial_var = "trial",
  time_var = "time_step",
  net_var = "A",
  ILVs_variable_effects = "",
  weibull_term = "",
  high_res = F
)
}
\arguments{
\item{transmission_func}{String (e.g. "standard", "freqdep_f", "freqdep_k")}

\item{is_distribution}{Boolean: Are edges drawn from posterior?}

\item{separate_s}{Boolean: Is s estimated separately per network?}

\item{veff_ID}{Character vector: Which params vary by ID}

\item{num_networks}{Integer: Number of networks}

\item{s_var}{String to use for s variable: e.g. "s_prime"}

\item{id_var}{String for id indexes: e.g., "id"}

\item{trial_var}{String for trial indexes: e.g., "trial"}

\item{time_var}{String: e.g., "time_step"}

\item{net_var}{String: e.g., "A"}

\item{ILVs_variable_effects}{String containing ILV effects on s}

\item{weibull_term}{String containing gamma etc}

\item{high_res}{boolean indicating if high_res}
}
\value{
String of Stan code for GQ block to accumulate psoc and psocn
}
\description{
Helper function that generates Stan code to compute P(e via ST) for \%ST calculation
}
