% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_networks.R
\name{grid_networks}
\alias{grid_networks}
\title{grid_networks()}
\usage{
grid_networks(event_data, networks)
}
\arguments{
\item{event_data}{event df}

\item{networks}{network df}
}
\value{
network dataframe
}
\description{
Helper function that completes network df with all dyad & trial combinations.
}
