% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_NBDA_STb.R
\name{import_NBDA_STb}
\alias{import_NBDA_STb}
\title{import_NBDA_STb: create STbayes_data object from nbda object}
\usage{
import_NBDA_STb(
  nbda_object,
  network_names = c("default"),
  ILVi = NULL,
  ILVs = NULL,
  ILVm = NULL
)
}
\arguments{
\item{network_names}{character vector of descriptive names of networks you're importing.}

\item{ILVi}{Optional character vector of ILVS to be considered when estimating asocial learning rate. If not specified, taken from NBDA object.}

\item{ILVs}{Optional character vector of ILVS to be considered when estimating social learning rate. If not specified, taken from NBDA object.}

\item{ILVm}{Optional character vector of ILVS to be considered in a multiplicative model. If not specified, taken from NBDA object.}

\item{diffusion_data}{dataframe with columns id, trial, time, max_time}

\item{networks}{dataframe with columns trial, from, to, and one or more columns of edge weights named descriptively. Optionally an integer time column can be provided for dynamic network analysis, although networks must be provided for each time period between transmission events.}

\item{ILV_metadata}{dataframe with columns id, and any individual-level variables that might be of interest}
}
\value{
A list containing properly formatted data to run social transmission models.
}
\description{
import_NBDA_STb: create STbayes_data object from nbda object
}
\examples{
nbda_object<-nbdaData_cTADA <- STbayes::nbdaData_cTADA
data_list <- import_NBDA_STb(nbda_object, network_names="assoc")
}
