% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_process.R
\name{OLD_process_networks_x_weights_hires}
\alias{OLD_process_networks_x_weights_hires}
\title{Helper function for preprocessing "high resolution" data}
\usage{
OLD_process_networks_x_weights_hires(event_data, t_weights, networks, D_data)
}
\arguments{
\item{event_data}{}
}
\description{
Helper function for preprocessing "high resolution" data
}
