% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_transformed_params_block.R
\name{optimize_transformed_params}
\alias{optimize_transformed_params}
\title{optimize_transformed_params: helper function to get rid of duplicate loops}
\usage{
optimize_transformed_params(transformed_params)
}
\arguments{
\item{transformed_params}{vector of strings}
}
\value{
string
}
\description{
optimize_transformed_params: helper function to get rid of duplicate loops
}
