% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_params.R
\name{order_params}
\alias{order_params}
\title{Order params}
\usage{
order_params(p)
}
\arguments{
\item{p}{name of parameter}
}
\value{
integer indicating order
}
\description{
Helper function for ordering params for STb_summary()
}
