% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LFO_helpers.R
\name{plot_ks}
\alias{plot_ks}
\title{plot_ks()}
\usage{
plot_ks(ks, L, k_thres = 0.7)
}
\arguments{
\item{ks}{Numeric vector of Pareto-\eqn{k} values from
\code{attr(x, "pointwise")$pareto_k} for an object returned by \code{\link[=STb_lfo]{STb_lfo()}}.}

\item{L}{Integer. Initial number of observations used before LFO prediction
begins.}

\item{k_thres}{Numeric. Pareto-\eqn{k} threshold to draw as a reference line.
Defaults to \code{0.7}.}
}
\description{
Create diagnostic plot of Pareto-\eqn{k} values from LFO-CV.
}
