% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_hires_complex_k.R
\name{process_hires_complex_k}
\alias{process_hires_complex_k}
\title{Title}
\usage{
process_hires_complex_k(event_data, networks, t_weights, D_data)
}
\arguments{
\item{transmission_type}{#' @importFrom data.table as.data.table setkey fifelse}
}
\description{
Title
}
