% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_duplicate_columns.R
\name{remove_duplicate_columns}
\alias{remove_duplicate_columns}
\title{remove_duplicate_columns()}
\usage{
remove_duplicate_columns(df)
}
\arguments{
\item{df}{a dataframe with possibly duplicated columns}
}
\value{
dataframe
}
\description{
Helper function to remove duplicated rows
}
