% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_ids.R
\name{standardize_ids}
\alias{standardize_ids}
\title{standardize_ids()}
\usage{
standardize_ids(
  networks,
  event_data,
  ILV_c = NULL,
  ILV_tv = NULL,
  t_weights = NULL
)
}
\arguments{
\item{networks}{networks dataframe with columns trial, from, to, and one or more columns of edge weights named descriptively.}

\item{event_data}{dataframe with columns id, trial, time, t_end}

\item{ILV_c}{optional dataframe with columns id, and any constant individual-level variables that might be of interest}

\item{ILV_tv}{optional dataframe with columns trial, id, time and any time-varying variables.}

\item{t_weights}{Optional dataframe with columns trial, id, time and t_weight.}
}
\value{
list of dataframes with filled out ids
}
\description{
Helper function that maps IDs across data with networks as reference
}
