% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STb_compare.R
\name{STb_compare}
\alias{STb_compare}
\title{STb_compare()}
\usage{
STb_compare(..., model_names = NULL, method = "loo-psis")
}
\arguments{
\item{...}{CmdStanMCMC model fits for comparison}

\item{model_names}{an optional list of model names, otherwise taken from object names}

\item{method}{a string either "loo-psis" or "waic" to indicate the method used for elpd. Defaults to "loo-psis". Pareto diagnostics are not calculated for WAIC.}
}
\value{
list containing loo_objects, comparison, and pareto_diagnostics if using loo-psis.
}
\description{
Convenience function for model comparison using either loo-psis or waic methods.
Returns list with relevant output from loo functions.
}
\examples{
data_list <- import_user_STb(STbayes::event_data, STbayes::edge_list)
model_full <- generate_STb_model(data_list, gq = TRUE, model_type = "full")
fit_full <- fit_STb(data_list,
    model_full,
    parallel_chains = 4,
    chains = 4,
    cores = 4,
    iter = 4000,
    refresh = 2000
)
# fit asocial model
model_asoc <- generate_STb_model(data_list, gq = TRUE, model_type = "asocial")
fit_asoc <- fit_STb(data_list,
    model_asoc,
    parallel_chains = 4,
    chains = 4,
    cores = 4,
    iter = 4000,
    refresh = 2000
)
mc <- STb_compare(fit_full, fit_asoc, method = "loo-psis")
# view comparison table
print(mc$comparison)

}
