% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STb_save.R
\name{STb_save}
\alias{STb_save}
\title{Save STbayes objects easily}
\usage{
STb_save(fit, output_dir = "cmdstan_saves", name = NULL)
}
\arguments{
\item{fit}{model fitted by STbayes}

\item{output_dir}{where do you want to save it???}

\item{name}{optional name to use for model}
}
\description{
Save STbayes objects easily
}
