% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_integer.R
\name{check_integer}
\alias{check_integer}
\title{check_integer: private function to check whether or not number is integer or integer-like. can't believe there's not a base function for this!}
\usage{
check_integer(x)
}
\arguments{
\item{x}{number or vector}
}
\value{
boolean values of whether a number is integer or integer-like
}
\description{
check_integer: private function to check whether or not number is integer or integer-like. can't believe there's not a base function for this!
}
