% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_ILV_datatype.R
\name{detect_ILV_datatype}
\alias{detect_ILV_datatype}
\title{Detect ILV datatype}
\usage{
detect_ILV_datatype(x)
}
\arguments{
\item{x}{column or vector of data}
}
\value{
string with category of variable
}
\description{
Helper function for importing ILVs
}
