% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_acqTime.R
\name{extract_acqTime}
\alias{extract_acqTime}
\title{Function to extract estimated acquisition times}
\usage{
extract_acqTime(fit, data_list, var_name = "acquisition_time")
}
\arguments{
\item{fit}{rstan fit object}

\item{data_list}{STb_data object}

\item{var_name}{name of where acquisition time is stored in rstan fit}
}
\value{
dataframe of ids, estimated timestep + HPD levels and observed timestep
}
\description{
Function to extract estimated acquisition times
}
