% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_acqTime.R
\name{extract_acqTime}
\alias{extract_acqTime}
\title{Function to extract estimated acquisition times}
\usage{
extract_acqTime(fit, data_list, prob = 0.95, var_name = "acquisition_time")
}
\arguments{
\item{fit}{CmdStanMCMC fit object}

\item{data_list}{STb_data object (must include ind_id and time)}

\item{prob}{interval for CI. defaults to .95}

\item{var_name}{variable name of acquisition time from GQ block (default = "acquisition_time")}
}
\value{
Dataframe of trial, individual, id, observed_time, mean/median/HPD of estimated time
}
\description{
Function to extract estimated acquisition times
}
