% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_LOOIC.R
\name{extract_LOOIC}
\alias{extract_LOOIC}
\title{Convenience to extract LOOIC scores}
\usage{
extract_LOOIC(fit)
}
\arguments{
\item{fit}{A STb model fit}
}
\value{
LOOIC estimate
}
\description{
Convenience to extract LOOIC scores
}
