% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_WAIC.R
\name{extract_WAIC}
\alias{extract_WAIC}
\title{Convenience to extract WAIC scores}
\usage{
extract_WAIC(fit)
}
\arguments{
\item{fit}{A STb model fit}
}
\value{
Table with WAIC estimates
}
\description{
Convenience to extract WAIC scores
}
