% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_STb_model.R
\name{fit_STb}
\alias{fit_STb}
\title{fit_STb()}
\usage{
fit_STb(data_list, model_obj, ...)
}
\arguments{
\item{data_list}{list object exported from import_user_STb or import_NBDA_STb}

\item{model_obj}{Can be either model object exported from generate_STb_model or a filename}

\item{...}{Additional arguments passed to mod$sample (e.g., chains, iter_warmup, iter_sampling, adapt_delta)}
}
\value{
A CmdStanMCMC fit object
}
\description{
Third step of analysis pipeline: Fits model using cmdstanr.
After fitting, consider using \code{\link[=STb_save]{STb_save()}} to store the output.
}
\examples{
data_list <- import_user_STb(STbayes::event_data, STbayes::edge_list)
model_obj <- generate_STb_model(data_list)
fit <- fit_STb(data_list,
    model_obj,
    parallel_chains = 4,
    chains = 4,
    cores = 4,
    iter = 4000,
    refresh = 1000
)
}
