% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_STb_model.R
\name{fit_STb}
\alias{fit_STb}
\title{Title}
\usage{
fit_STb(
  data_list,
  model_obj,
  chains = 1,
  cores = 1,
  iter = 1000,
  control = list(),
  algorithm = "NUTS"
)
}
\arguments{
\item{data_list}{list object exported from import_user_STb or import_NBDA_STb}

\item{model_obj}{Can be either model object exported from generate_STb_model or filename}

\item{chains}{Integer Number of chains to run (default=1)}

\item{cores}{Integer Number of cores to use (default=1)}

\item{iter}{Integer Number of iterations to run for}

\item{control}{List of arguments to pass to control.}

\item{algorithm}{Defaults to "NUTS". If running asocial OADA, specify "Fixed_param".}
}
\value{
rstan fit
}
\description{
Title
}
\examples{
data_list_user = import_user_STb(diffusion_data, edge_list)
model_obj = generate_STb_model(data_list_user)
fit = fit_STb(data_list_user, model_obj, chains = 5, cores = 5, iter=2000, control = list(adapt_delta=0.99))
#or alternatively
fit = fit_STb(data_list_user, "path/to/your/model.stan", chains = 5, cores = 5, iter=2000, control = list(adapt_delta=0.99))
}
