% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_STb_model.R
\name{generate_STb_model}
\alias{generate_STb_model}
\title{Dynamically generate STAN model based on input data}
\usage{
generate_STb_model(STb_data, N_veff = 0, gq = TRUE, est_acqTime = FALSE)
}
\arguments{
\item{STb_data}{a list of formatted data returned from the STbayes_data() function}

\item{N_veff}{Number of varying effects. Defaults to 2 (one for strength of SL, one for asocial rate)}
}
\value{
A STAN model (character) that is customized to the input data.
}
\description{
Dynamically generate STAN model based on input data
}
\examples{
#very mock data
diffusion_data <- data.frame(
  id = c("A", "B", "C", "D", "E", "F"),
  trial = c(1, 1, 1, 2, 2, 2),
  time = c(0, 1, 2, 0, 1, 4),
  max_time = c(3, 3, 3, 4, 4, 4)
)
networks <- data.frame(
  trial = c(1, 1, 1, 2, 2, 2),
  from = c("A", "A", "B", "D", "D", "E"),
  to = c("B", "C", "C", "E", "F", "F"),
  kin = c(1, 0, 1, 0, 1, 1),
  inverse_distance = c(0, 1, .5, .25, .1, 0)
)
ILV_metadata <- data.frame(
  id = c("A", "B", "C", "D", "E", "F"),
  age = c(2, 3, 4, 2, 5, 6),
  sex = c(0, 1, 1, 0, 1, 0) # Factor ILVs must be input as numeric
)
data_list <- import_user_STb(
  diffusion_data = diffusion_data,
  networks = networks,
  ILV_metadata = ILV_metadata,
  ILVi = c("age"), # Use only 'age' for asocial learning
  ILVs = c("sex") # Use only 'sex' for social learning
)

model = generate_STb_model(data_list)
model = generate_STb_model(data_list, N_veff=2) # estimate varying effects for s and lambda_0.
print(model)
}
