% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_STb_model_OADA.R
\name{generate_STb_model_OADA}
\alias{generate_STb_model_OADA}
\title{generate_STb_model_OADA()}
\usage{
generate_STb_model_OADA(
  STb_data,
  model_type = "full",
  transmission_func = "standard",
  veff_ID = c(),
  gq = TRUE,
  priors = list()
)
}
\arguments{
\item{STb_data}{a list of formatted data returned from the STbayes_data() function}

\item{model_type}{string specifying the model type: "asocial" or "full"}

\item{transmission_func}{string specifying transmission function: "standard", "freqdep_f" or "freqdep_k" for frequency dependent complex contagion. Defaults to "standard".}

\item{veff_ID}{Parameters for which to estimate varying effects by individuals. Default is no varying effects.}

\item{gq}{Boolean to indicate whether the generated quantities block is added (incl. ll for WAIC)}

\item{priors}{named list with strings containing the prior for log s or f. defaults to list(log_s = "uniform(-10, 10)", log_f = "normal(0,1)")}
}
\value{
A STAN model (character) that is customized to the input data.
}
\description{
Helper function that creates stan code for OADA type models.
}
