% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_STb_model_OADA.R
\name{generate_STb_model_OADA}
\alias{generate_STb_model_OADA}
\title{Dynamically generate STAN model based on input data}
\usage{
generate_STb_model_OADA(
  STb_data,
  model_type = "full",
  transmission_func = "standard",
  veff_ID = c(),
  gq = TRUE,
  priors = list(log_s = "uniform(-10,10)", log_f = "normal(0,1)")
)
}
\arguments{
\item{STb_data}{a list of formatted data returned from the STbayes_data() function}

\item{model_type}{string specifying the model type: "asocial" or "full"}

\item{transmission_func}{string specifying transmission function: "standard", "freq-dep" for frequency dependent complex contagion. Defaults to "standard".}

\item{veff_ID}{Parameters for which to estimate varying effects by individuals. Default is no varying effects.}

\item{gq}{Boolean to indicate whether the generated quantities block is added (incl. ll for WAIC)}

\item{priors}{named list with strings containing the prior for log s or f. defaults to list(log_s = "uniform(-10, 10)", log_f = "normal(0,1)")}
}
\value{
A STAN model (character) that is customized to the input data.
}
\description{
Dynamically generate STAN model based on input data
}
\examples{
#very mock data
event_data <- data.frame(
  id = c("A", "B", "C", "D", "E", "F"),
  trial = c(1, 1, 1, 2, 2, 2),
  time = c(0, 1, 2, 0, 1, 4),
  max_time = c(3, 3, 3, 4, 4, 4)
)
networks <- data.frame(
  trial = c(1, 1, 1, 2, 2, 2),
  from = c("A", "A", "B", "D", "D", "E"),
  to = c("B", "C", "C", "E", "F", "F"),
  kin = c(1, 0, 1, 0, 1, 1),
  inverse_distance = c(0, 1, .5, .25, .1, 0)
)
ILV_metadata <- data.frame(
  id = c("A", "B", "C", "D", "E", "F"),
  age = c(2, 3, 4, 2, 5, 6),
  sex = c(0, 1, 1, 0, 1, 0), # Factor ILVs must be input as numeric
  weight = c(0.5, .25, .3, 0, -.2, -.4)
)
data_list <- import_user_STb(
  event_data = event_data,
  networks = networks,
  ILV_metadata = ILV_metadata,
  ILVi = c("age"), # Use only 'age' for asocial learning
  ILVs = c("sex"), # Use only 'sex' for social learning
  ILVm = c("weight") # Use weight for multiplicative effect on asocial and social learning
)

model = generate_STb_model_OADA(data_list) # no varying effects
model = generate_STb_model_OADA(data_list, veff_ID = c("s")) # estimate varying effects by ID for strength of social transmission. Baseline is not estimated for OADA type models.
print(model)

}
