% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_network_term.R
\name{get_network_term}
\alias{get_network_term}
\title{get_network_term()}
\usage{
get_network_term(
  transmission_func = "standard",
  is_distribution = FALSE,
  num_networks = 1,
  veff_params = c(),
  s_var = "s_prime",
  net_var = "A",
  net_index = "network",
  id_var = "id",
  veff_type = c(),
  veff_idx = "id",
  trial_var = "trial",
  time_var = "time_step",
  high_res = F
)
}
\arguments{
\item{transmission_func}{simple or complex transmission}

\item{is_distribution}{did the user supply a posterior distribution for edges}

\item{num_networks}{integer of number of networks}

\item{veff_params}{vector of varying effects}

\item{s_var}{string used to represent s term ("s_prime" or "s_direct")}

\item{net_var}{string used to represent network variable "A"}

\item{net_index}{string used to index networks "network"}

\item{id_var}{string used to index individuals. "id" for tada, need to use "id" and "j" for oada}

\item{veff_type}{string or character vector of whether veffs are id, trial or both}

\item{veff_idx}{string used to index parameters for varying effects}

\item{trial_var}{string used to index trial "trial"}

\item{time_var}{string used to index time "time_step"}

\item{high_res}{boolean indicating if high res}
}
\value{
string of stan code to be used in the model for calculating network effects
}
\description{
Helper function to generate network terms for STAN
}
