% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_NBDA_STb.R
\name{import_NBDA_STb}
\alias{import_NBDA_STb}
\title{import_NBDA_STb: create STbayes_data object from nbda object}
\usage{
import_NBDA_STb(
  nbda_object,
  network_names = c("default"),
  multinetwork_s = c("separate", "shared"),
  ILVi = NULL,
  ILVs = NULL,
  ILVm = NULL
)
}
\arguments{
\item{nbda_object}{object of NBDAdata class}

\item{network_names}{character vector of descriptive names of networks you're importing.}

\item{multinetwork_s}{"separate" or "shared". If supplying more than one network, specify whether each network receives it's own s parameter, or shares a single parameter.}

\item{ILVi}{Optional character vector of ILVS to be considered when estimating intrinsic rate. If not specified, taken from NBDA object.}

\item{ILVs}{Optional character vector of ILVS to be considered when estimating social transmission rate. If not specified, taken from NBDA object.}

\item{ILVm}{Optional character vector of ILVS to be considered in a multiplicative model. If not specified, taken from NBDA object.}
}
\value{
A list containing properly formatted data to run social transmission models.
}
\description{
import_NBDA_STb: create STbayes_data object from nbda object
}
\examples{
nbda_object <- nbdaData_cTADA <- STbayes::nbdaData_cTADA
data_list <- import_NBDA_STb(nbda_object, network_names = "assoc")
}
