% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_user_STb.R
\name{import_user_STb}
\alias{import_user_STb}
\title{import_user_STb: Create STbayes_data object from user supplied data}
\usage{
import_user_STb(
  diffusion_data,
  networks,
  ILV_c = NULL,
  ILV_tv = NULL,
  ILVi = NULL,
  ILVs = NULL,
  ILVm = NULL
)
}
\arguments{
\item{diffusion_data}{dataframe with columns id, trial, time, max_time}

\item{networks}{dataframe with columns trial, from, to, and one or more columns of edge weights named descriptively. Optionally an integer time column can be provided for dynamic network analysis, although networks must be provided for each time period between transmission events.}

\item{ILV_c}{optional dataframe with columns id, and any individual-level variables that might be of interest}

\item{ILV_tv}{optional dataframe with columns trial, id, time and any time-varying variables. Variable values should summarize the variable for each inter-acquisition period.}

\item{ILVi}{Optional character vector of column names from ILV metadata to be considered when estimating asocial learning rate. If not specified, all ILV are applied to both.}

\item{ILVs}{Optional character vector of column names from ILV metadata to be considered when estimating social learning rate. If not specified, all ILV are applied to both.}

\item{ILVm}{Optional character vector of column names from ILV metadata to be considered in a multiplicative model.}
}
\value{
A list object containing properly formatted data to run social transmission models.
}
\description{
import_user_STb: Create STbayes_data object from user supplied data
}
\examples{
# very mock data
diffusion_data <- data.frame(
  trial = rep(1:2, each = 3),
  id = LETTERS[1:6],
  time = c(0, 1, 2, 0, 1, 4),
  max_time = c(3, 3, 3, 4, 4, 4)
)
networks <- data.frame(
  trial = rep(1:2, each = 3),
  from = c("A", "A", "B", "D", "D", "E"),
  to = c("B", "C", "C", "E", "F", "F"),
  kin = c(1, 0, 1, 0, 1, 1),
  inverse_distance = c(0, 1, .5, .25, .1, 0)
)
ILV_c <- data.frame(
  id = LETTERS[1:6],
  age = c(-1, -2, 0, 1, 2), # continuous variables should be normalized
  sex = c(0, 1, 1, 0, 1, 0), # Factor ILVs must be input as numeric
  weight = c(0.5, .25, .3, 0, -.2, -.4)
)
ILV_tv <- data.frame(
  trial = c(rep(1, each = 9), rep(2, each = 9)),
  id = c(rep(LETTERS[1:3], each = 3), rep(LETTERS[4:6], each = 3)),
  # these times correspond to the inter-acquisition periods
  # e.g. 1 is from [t_0 to t_1), 2 is [t_1 to t_2), 3 = [t_2 to t_3 or t_end] if censored inds. present)
  time = c(rep(1:3, times = 3), rep(1:3, times = 3)),
  # ensure the variable is summarizing these inter-acquisition time periods
  dist_from_task = rnorm(18)
)
imported_data <- import_user_STb(
  diffusion_data = diffusion_data,
  networks = networks,
  ILV_c = ILV_c,
  ILV_tv = ILV_tv,
  ILVi = c("age", "dist_from_task"), # Use 'age' and time-varying ILV 'dist_from_task' for asocial learning
  ILVs = c("sex"), # Use only 'sex' for social learning
  ILVm = c("weight") # Use weight for multiplicative effect on asocial and social learning
)

}
