% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_ILVs_for_generate_model.R
\name{process_ILVs}
\alias{process_ILVs}
\title{Helper function for generating ctada}
\usage{
process_ILVs(
  ilv_vars,
  ilv_vars_clean,
  veff_ID,
  suffix,
  STb_data,
  count_start,
  prior_beta
)
}
\arguments{
\item{count_start}{}
}
\description{
Helper function for generating ctada
}
