% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_networks_x_weights_hires.R
\name{process_networks_x_weights_hires}
\alias{process_networks_x_weights_hires}
\title{Helper function for preprocessing "high resolution" data}
\usage{
process_networks_x_weights_hires(event_data, t_weights, networks, D_data)
}
\arguments{
\item{event_data}{}
}
\description{
Helper function for preprocessing "high resolution" data
}
