% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_N_veff.R
\name{return_N_veff}
\alias{return_N_veff}
\title{return_N_veff()}
\usage{
return_N_veff(text)
}
\arguments{
\item{text}{model object (text)}
}
\value{
integer of number of varying effects present in model
}
\description{
Helper function to find the number of parameters with varying effects
}
