% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STb_save.R
\name{STb_save}
\alias{STb_save}
\title{Save STbayes objects easily}
\usage{
STb_save(fit, name = NULL, path = "cmdstan_saves")
}
\arguments{
\item{fit}{model fitted by STbayes}

\item{name}{optional name to use for model}

\item{path}{where do you want to save it???}
}
\description{
Save STbayes objects easily
}
