% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STb_summary.R
\name{STb_summary}
\alias{STb_summary}
\title{Return summary table for CmdStanR fit}
\usage{
STb_summary(
  fit,
  depth = 1,
  prob = 0.95,
  ignore_params = c("lp__", "idx", "log_lik", "log_lik_matrix", "acquisition_time",
    "z_ID", "Rho_ID", "v_ID", ".chain", ".iteration", ".draw"),
  digits = 3
)
}
\arguments{
\item{fit}{CmdStanMCMC model fit}

\item{depth}{integer depth of multidimensional parameters to extract}

\item{prob}{double limits for HPD of estimates (default = 0.95)}

\item{ignore_params}{character vector of parameters to ignore}

\item{digits}{integer of digits to round to}
}
\value{
Summary table
}
\description{
Return summary table for CmdStanR fit
}
